chore: move anvil widgets to wds (#35812)
## Description
- Moved Section and Zone widgets to wds folder

Fixes #35810

> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
