+++ app/client/src/layoutSystems/anvil/integrations/sagas/anvilDraggingSagas/anvilDraggingSagas.test.ts
@@ -6,8 +6,8 @@ import { LayoutComponentTypes } from "layoutSystems/anvil/utils/anvilTypes";
 import { expectSaga } from "redux-saga-test-plan";
 import { getWidgets } from "sagas/selectors";
 import { registerWidgets } from "WidgetProvider/factory/registrationHelper";
-import { SectionWidget } from "widgets/anvil/SectionWidget";
-import { ZoneWidget } from "widgets/anvil/ZoneWidget";
+import { WDSSectionWidget } from "widgets/wds/WDSSectionWidget";
+import { WDSZoneWidget } from "widgets/wds/WDSZoneWidget";
 import { WDSButtonWidget } from "widgets/wds/WDSButtonWidget";
 import {
   getCanvasWidth,
@@ -39,8 +39,8 @@ describe("", () => {
   beforeAll(() => {
     registerLayoutComponents();
     registerWidgets([
-      SectionWidget,
-      ZoneWidget,
+      WDSSectionWidget,
+      WDSZoneWidget,
       WDSButtonWidget,
       WDSModalWidget,
     ]);
