+++ app/client/src/layoutSystems/anvil/utils/constants.ts
@@ -1,6 +1,6 @@
 import { FlexLayerAlignment } from "layoutSystems/common/utils/constants";
 import type { HighlightPayload, HighlightRenderInfo } from "./anvilTypes";
-import { anvilWidgets } from "widgets/anvil/constants";
+import { anvilWidgets } from "widgets/wds/constants";
 
 export const MOBILE_BREAKPOINT = 480;
