+++ app/client/src/layoutSystems/anvil/utils/paste/sectionPasteUtils.ts
@@ -6,7 +6,7 @@ import type {
 } from "./types";
 import { getDestinedParent } from "./destinationUtils";
 import type { FlattenedWidgetProps } from "WidgetProvider/constants";
-import { anvilWidgets } from "widgets/anvil/constants";
+import { anvilWidgets } from "widgets/wds/constants";
 import type { LayoutProps } from "../anvilTypes";
 import { all, call } from "redux-saga/effects";
 import { addPastedWidgets } from "./utils";
