--- app/client/src/widgets/anvil/SectionWidget/index.tsx
@@ -1,3 +1,0 @@
-import SectionWidget from "./widget";
-
-export { SectionWidget };
