--- app/client/src/widgets/anvil/ZoneWidget/index.tsx
@@ -1,3 +1,0 @@
-import ZoneWidget from "./widget";
-
-export { ZoneWidget };
