+++ app/client/src/widgets/wds/WDSSectionWidget/config/baseConfig.ts
