+++ app/client/src/widgets/wds/WDSSectionWidget/config/defaultConfig.ts
