+++ app/client/src/widgets/wds/WDSSectionWidget/config/index.ts
