+++ app/client/src/widgets/wds/WDSSectionWidget/config/methodsConfig.ts
