+++ app/client/src/widgets/wds/WDSSectionWidget/config/propertyPaneStyle.ts
