+++ app/client/src/widgets/wds/WDSSectionWidget/index.tsx
@@ -1,0 +1,3 @@
+import WDSSectionWidget from "./widget";
+
+export { WDSSectionWidget };
