+++ app/client/src/widgets/wds/WDSStatsWidget/widget/index.tsx
@@ -7,8 +7,8 @@ import * as config from "../config";
 import { StatsComponent } from "../component";
 import type { StatsWidgetProps } from "./types";
 import type { AnvilConfig } from "WidgetProvider/constants";
-import { Elevations } from "widgets/anvil/constants";
-import { ContainerComponent } from "widgets/anvil/Container";
+import { Elevations } from "widgets/wds/constants";
+import { ContainerComponent } from "widgets/wds/Container";
 
 class WDSStatsWidget extends BaseWidget<StatsWidgetProps, WidgetState> {
   constructor(props: StatsWidgetProps) {
