+++ app/client/src/widgets/wds/WDSZoneWidget/index.tsx
@@ -1,0 +1,3 @@
+import WDSZoneWidget from "./widget";
+
+export { WDSZoneWidget };
