+++ app/client/src/widgets/wds/WDSZoneWidget/widget/config/baseConfig.ts
