+++ app/client/src/widgets/wds/WDSZoneWidget/widget/config/defaultConfig.ts
