+++ app/client/src/widgets/wds/WDSZoneWidget/widget/config/methodsConfig.ts
