+++ app/client/src/widgets/wds/WDSZoneWidget/widget/index.tsx
@@ -22,9 +22,9 @@ import BaseWidget from "widgets/BaseWidget";
 import type { WidgetProps, WidgetState } from "widgets/BaseWidget";
 import type { LayoutProps } from "layoutSystems/anvil/utils/anvilTypes";
 import type { ContainerWidgetProps } from "widgets/ContainerWidget/widget";
-import { ContainerComponent } from "widgets/anvil/Container";
+import { ContainerComponent } from "widgets/wds/Container";
 import { LayoutProvider } from "layoutSystems/anvil/layoutComponents/LayoutProvider";
-import { Elevations, anvilWidgets } from "widgets/anvil/constants";
+import { Elevations, anvilWidgets } from "widgets/wds/constants";
 import type { CanvasWidgetsReduxState } from "reducers/entityReducers/canvasWidgetsReducer";
 import type {
   CopiedWidgetData,
@@ -35,7 +35,7 @@ import { call } from "redux-saga/effects";
 import { pasteWidgetsInZone } from "layoutSystems/anvil/utils/paste/zonePasteUtils";
 import { SectionColumns } from "layoutSystems/anvil/sectionSpaceDistributor/constants";
 
-class ZoneWidget extends BaseWidget<ZoneWidgetProps, WidgetState> {
+class WDSZoneWidget extends BaseWidget<WDSZoneWidgetProps, WidgetState> {
   static type = anvilWidgets.ZONE_WIDGET;
 
   static getConfig(): WidgetBaseConfiguration {
@@ -153,8 +153,8 @@ class ZoneWidget extends BaseWidget<ZoneWidgetProps, WidgetState> {
   }
 }
 
-export interface ZoneWidgetProps extends ContainerWidgetProps<WidgetProps> {
+export interface WDSZoneWidgetProps extends ContainerWidgetProps<WidgetProps> {
   layout: LayoutProps[];
 }
 
-export default ZoneWidget;
+export default WDSZoneWidget;
